fix(rubocop): split TradeTariffBackend config readers
Structure-only: move ENV/config readers into themed Config::* modules
and include them from Config. Preserve defaults including
openai_model_pricing added on main after the design SHA.

Issue: No ticket/issue
