+++ app/lib/trade_tariff_backend/config/goods_nomenclature.rb
@@ -1,0 +1,9 @@
+module TradeTariffBackend
+  module Config
+    module GoodsNomenclature
+      def goods_nomenclature_label_page_size
+        ENV.fetch('GOODS_NOMENCLATURE_LABEL_PAGE_SIZE', '10').to_i
+      end
+    end
+  end
+end
