+++ app/lib/trade_tariff_backend/config/xe.rb
@@ -1,0 +1,17 @@
+module TradeTariffBackend
+  module Config
+    module XE
+      def xe_api_url
+        ENV.fetch('XE_API_URL', 'https://xecdapi.xe.com')
+      end
+
+      def xe_api_username
+        ENV['XE_API_USERNAME']
+      end
+
+      def xe_api_password
+        ENV['XE_API_PASSWORD']
+      end
+    end
+  end
+end
