+++ config/debride.whitelist
@@ -501,12 +501,12 @@ TariffSynchronizer::TariffDownloader#create_record_for_not_found_response
 TariffSynchronizer::TariffDownloader#create_record_for_successful_response
 TradeTariffBackend#configure
 TradeTariffBackend#data_migration_path
-TradeTariffBackend::Config#allow_missing_migration_files
-TradeTariffBackend::Config#max_threads
-TradeTariffBackend::Config#redis_config
-TradeTariffBackend::Config#slack_channel
-TradeTariffBackend::Config#slack_username
-TradeTariffBackend::Config#slack_web_hook_url
+TradeTariffBackend::Config::Infrastructure#allow_missing_migration_files
+TradeTariffBackend::Config::Infrastructure#max_threads
+TradeTariffBackend::Config::Redis#redis_config
+TradeTariffBackend::Config::Slack#slack_channel
+TradeTariffBackend::Config::Slack#slack_username
+TradeTariffBackend::Config::Slack#slack_web_hook_url
 TradeTariffBackend::SearchClient#delete_by_name
 TradeTariffBackend::SearchClient#index_by_name
 TradeTariffBackend::SearchResponse#hits
